+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * [Change] Better detect if any of the peers in the KAS connection got destroyed for any reason. The link gets broken in this case.
 * [Fix #302] GUI does not respect hide/show function.
 * [Fix #306] Logs spam from the parts dropped on the ground.
+* [Fix #307] Interactive attach mode conflicts with construction mode.
 * [Fix #308] Linked parts can be dragged in construct mode.
 * [Fix #314] Retract cable option is visible when the connector is locked.
 * [Fix #315] Attaching KAS links resets EVA editor parts highlighting.